mysql update 3
- finish fixing auth flow and ut auth
- other housekeeping
- fix for flexCarousel to adapt to other pattern changes
- finish converting data design and sample data

NEXT UP:
- query recipe data
- return data from recipe read route
