+++ client/components/genericWidgets/flexCarousel/flexCarousel.html
@@ -1,9 +1,9 @@
 <span class="flexCarousel">
   <span class="crslPrevWrapper">
-    <span class="crslPrev" onclick="crslCycle(event, this, 'prev')"></span>
+    <span class="crslPrev" onclick="cDI.components.flexCarousel.crslCycle(event, this, 'prev')"></span>
   </span>
   <span class="crslNextWrapper">
-    <span class="crslNext" onclick="crslCycle(event, this, 'next')"></span>
+    <span class="crslNext" onclick="cDI.components.flexCarousel.crslCycle(event, this, 'next')"></span>
   </span>
   <span class="crslSeatPanel"></span>
 </span>
