+++ client/pages/cookbook/cookbook.js
@@ -3,7 +3,8 @@ cDI.pages.cookbook = {
   init: async () => {
     await cDI.remote.asyncGetScript(`js/services/recipeService.js`)
     var recipes = await cDI.services.recipe.getAllRecipes()
-    var favorites = await cDI.components.recipeCard.buildRecipeCardList(recipes)
+    console.log()
+    // var favorites = await cDI.components.recipeCard.buildRecipeCardList(recipes)
     $("#counterTop").append(favorites)
     $("#counterTop").append("<span></span>")
   }
